+++ ignite/cmd/network_chain_publish.go
@@ -8,7 +8,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/tendermint/spn/pkg/chainid"
-	campaigntypes "github.com/tendermint/spn/x/campaign/types"
 
 	"github.com/ignite-hq/cli/ignite/pkg/cliui"
 	"github.com/ignite-hq/cli/ignite/pkg/cliui/icons"
@@ -41,6 +40,7 @@ func NewNetworkChainPublish() *cobra.Command {
 		RunE:  networkChainPublishHandler,
 	}
 
+	flagSetClearCache(c)
 	c.Flags().String(flagBranch, "", "Git branch to use for the repo")
 	c.Flags().String(flagTag, "", "Git tag to use for the repo")
 	c.Flags().String(flagHash, "", "Git hash to use for the repo")
@@ -85,6 +85,11 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 	source, err := xurl.MightHTTPS(args[0])
 	if err != nil {
 		return fmt.Errorf("invalid source url format: %w", err)
+	}
+
+	cacheStorage, err := newCache(cmd)
+	if err != nil {
+		return err
 	}
 
 	if campaign != 0 && campaignTotalSupplyStr != "" {
@@ -202,7 +207,7 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 			return err
 		}
 
-		publishOptions = append(publishOptions, network.WithShares(campaigntypes.NewSharesFromCoins(coins)))
+		publishOptions = append(publishOptions, network.WithPercentageShares(coins))
 	}
 
 	// init the chain.
@@ -213,7 +218,7 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 
 	if noCheck {
 		publishOptions = append(publishOptions, network.WithNoCheck())
-	} else if err := c.Init(cmd.Context()); err != nil { // initialize the chain for checking.
+	} else if err := c.Init(cmd.Context(), cacheStorage); err != nil { // initialize the chain for checking.
 		return err
 	}