+++ ignite/cmd/network_chain_show_genesis.go
@@ -19,6 +19,7 @@ func newNetworkChainShowGenesis() *cobra.Command {
 		RunE:  networkChainShowGenesisHandler,
 	}
 
+	flagSetClearCache(c)
 	c.Flags().String(flagOut, "./genesis.json", "Path to output Genesis file")
 
 	return c
@@ -29,6 +30,11 @@ func networkChainShowGenesisHandler(cmd *cobra.Command, args []string) error {
 	defer session.Cleanup()
 
 	out, _ := cmd.Flags().GetString(flagOut)
+	
+	cacheStorage, err := newCache(cmd)
+	if err != nil {
+		return err
+	}
 
 	nb, launchID, err := networkChainLaunch(cmd, args, session)
 	if err != nil {
@@ -63,12 +69,15 @@ func networkChainShowGenesisHandler(cmd *cobra.Command, args []string) error {
 		}
 
 		// create the chain in a temp dir
-		home := filepath.Join(os.TempDir(), "spn/temp", chainLaunch.ChainID)
-		defer os.RemoveAll(home)
+		tmpHome, err := os.MkdirTemp("", "*-spn")
+		if err != nil {
+			return err
+		}
+		defer os.RemoveAll(tmpHome)
 
-		c.SetHome(home)
+		c.SetHome(tmpHome)
 
-		err = c.Prepare(cmd.Context(), genesisInformation)
+		err = c.Prepare(cmd.Context(), cacheStorage, genesisInformation)
 		if err != nil {
 			return err
 		}