+++ ignite/pkg/cosmosgen/generate_javascript.go
@@ -2,6 +2,7 @@ package cosmosgen
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
@@ -9,8 +10,9 @@ import (
 	"github.com/iancoleman/strcase"
 	"golang.org/x/sync/errgroup"
 
+	"github.com/ignite-hq/cli/ignite/pkg/cache"
 	"github.com/ignite-hq/cli/ignite/pkg/cosmosanalysis/module"
-	"github.com/ignite-hq/cli/ignite/pkg/giturl"
+	"github.com/ignite-hq/cli/ignite/pkg/dirchange"
 	"github.com/ignite-hq/cli/ignite/pkg/gomodulepath"
 	"github.com/ignite-hq/cli/ignite/pkg/localfs"
 	"github.com/ignite-hq/cli/ignite/pkg/nodetime/programs/sta"
@@ -29,7 +31,10 @@ var (
 	}
 )
 
-const vuexRootMarker = "vuex-root"
+const (
+	vuexRootMarker          = "vuex-root"
+	dirchangeCacheNamespace = "generate.javascript.dirchange"
+)
 
 type jsGenerator struct {
 	g *generator
@@ -60,10 +65,28 @@ func (g *jsGenerator) generateModules() error {
 
 	gg := &errgroup.Group{}
 
+	dirCache := cache.New[[]byte](g.g.cacheStorage, dirchangeCacheNamespace)
 	add := func(sourcePath string, modules []module.Module) {
 		for _, m := range modules {
 			m := m
-			gg.Go(func() error { return g.generateModule(g.g.ctx, tsprotoPluginPath, sourcePath, m) })
+			gg.Go(func() error {
+				cacheKey := m.Pkg.Path
+				paths := append([]string{m.Pkg.Path, g.g.o.jsOut(m)}, g.g.o.includeDirs...)
+				changed, err := dirchange.HasDirChecksumChanged(dirCache, cacheKey, sourcePath, paths...)
+				if err != nil {
+					return err
+				}
+
+				if !changed {
+					return nil
+				}
+
+				if err := g.generateModule(g.g.ctx, tsprotoPluginPath, sourcePath, m); err != nil {
+					return err
+				}
+
+				return dirchange.SaveDirChecksum(dirCache, cacheKey, sourcePath, paths...)
+			})
 		}
 	}
 
@@ -164,10 +187,7 @@ func (g *jsGenerator) generateVuexModuleLoader() error {
 		return err
 	}
 
-	chainURL, err := giturl.Parse(chainPath.RawPath)
-	if err != nil {
-		return err
-	}
+	appModulePath := gomodulepath.ExtractAppPath(chainPath.RawPath)
 
 	type module struct {
 		Name     string
@@ -177,12 +197,10 @@ func (g *jsGenerator) generateVuexModuleLoader() error {
 	}
 
 	data := struct {
-		Modules []module
-		User    string
-		Repo    string
+		Modules     []module
+		PackageName string
 	}{
-		User: chainURL.User,
-		Repo: chainURL.Repo,
+		PackageName: fmt.Sprintf("%s-js", strings.ReplaceAll(appModulePath, "/", "-")),
 	}
 
 	for _, path := range modulePaths {