+++ ignite/pkg/relayer/config/config.go
@@ -58,6 +58,7 @@ type Chain struct {
 	RPCAddress    string `json:"rpc_address" yaml:"rpc_address"`
 	GasPrice      string `json:"gas_price" yaml:"gas_price,omitempty"`
 	GasLimit      int64  `json:"gas_limit" yaml:"gas_limit,omitempty"`
+	ClientID      string `json:"client_id" yaml:"client_id,omitempty"`
 }
 
 type Path struct {
@@ -91,4 +92,8 @@ func Get() (Config, error) {
 func Save(c Config) error {
 	c.Version = supportVersion
 	return confile.New(confile.DefaultYAMLEncodingCreator, configPath).Save(c)
+}
+
+func Delete() error {
+	return os.RemoveAll(configPath)
 }