+++ ignite/services/chain/serve.go
@@ -17,6 +17,7 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/ignite-hq/cli/ignite/chainconfig"
+	"github.com/ignite-hq/cli/ignite/pkg/cache"
 	chaincmdrunner "github.com/ignite-hq/cli/ignite/pkg/chaincmd/runner"
 	"github.com/ignite-hq/cli/ignite/pkg/cosmosfaucet"
 	"github.com/ignite-hq/cli/ignite/pkg/dirchange"
@@ -31,14 +32,17 @@ const (
 	// exportedGenesis is the name of the exported genesis file for a chain
 	exportedGenesis = "exported_genesis.json"
 
-	// sourceChecksum is the file containing the checksum to detect source modification
-	sourceChecksum = "source_checksum.txt"
+	// sourceChecksumKey is the cache key for the checksum to detect source modification
+	sourceChecksumKey = "source_checksum"
 
-	// binaryChecksum is the file containing the checksum to detect binary modification
-	binaryChecksum = "binary_checksum.txt"
+	// binaryChecksumKey is the cache key for the checksum to detect binary modification
+	binaryChecksumKey = "binary_checksum"
 
-	// configChecksum is the file containing the checksum to detect config modification
-	configChecksum = "config_checksum.txt"
+	// configChecksumKey is the cache key for containing the checksum to detect config modification
+	configChecksumKey = "config_checksum"
+
+	// serveDirchangeCacheNamespace is the name of the cache namespace for detecting changes in directories
+	serveDirchangeCacheNamespace = "serve.dirchange"
 )
 
 var (
@@ -82,7 +86,7 @@ func ServeResetOnce() ServeOption {
 }
 
 // Serve serves an app.
-func (c *Chain) Serve(ctx context.Context, options ...ServeOption) error {
+func (c *Chain) Serve(ctx context.Context, cacheStorage cache.Storage, options ...ServeOption) error {
 	serveOptions := newServeOption()
 
 	// apply the options
@@ -136,7 +140,7 @@ func (c *Chain) Serve(ctx context.Context, options ...ServeOption) error {
 				shouldReset := serveOptions.forceReset || serveOptions.resetOnce
 
 				// serve the app.
-				err = c.serve(serveCtx, shouldReset)
+				err = c.serve(serveCtx, cacheStorage, shouldReset)
 				serveOptions.resetOnce = false
 
 				switch {
@@ -180,8 +184,8 @@ func (c *Chain) Serve(ctx context.Context, options ...ServeOption) error {
 						c.ev.Send(
 							fmt.Sprintf(
 								"%s %s",
-								colors.Info("Blockchain failed to start.If the new code is no longer compatible with the saved state, you can reset the database by launching:"),
-								"starport chain serve --reset-once",
+								colors.Info("Blockchain failed to start.\nIf the new code is no longer compatible with the saved state, you can reset the database by launching:"),
+								"ignite chain serve --reset-once",
 							),
 						)
 						return fmt.Errorf("cannot run %s", startErr.AppName)
@@ -240,6 +244,7 @@ func (c *Chain) watchAppBackend(ctx context.Context) error {
 		localfs.WatcherWorkdir(c.app.Path),
 		localfs.WatcherOnChange(c.refreshServe),
 		localfs.WatcherIgnoreHidden(),
+		localfs.WatcherIgnoreFolders(),
 		localfs.WatcherIgnoreExt(ignoredExts...),
 	)
 }
@@ -247,7 +252,7 @@ func (c *Chain) watchAppBackend(ctx context.Context) error {
 // serve performs the operations to serve the blockchain: build, init and start
 // if the chain is already initialized and the file didn't changed, the app is directly started
 // if the files changed, the state is imported
-func (c *Chain) serve(ctx context.Context, forceReset bool) error {
+func (c *Chain) serve(ctx context.Context, cacheStorage cache.Storage, forceReset bool) error {
 	conf, err := c.Config()
 	if err != nil {
 		return &CannotBuildAppError{err}
@@ -258,13 +263,10 @@ func (c *Chain) serve(ctx context.Context, forceReset bool) error {
 		return err
 	}
 
-	saveDir, err := c.chainSavePath()
-	if err != nil {
-		return err
-	}
-
 	// isInit determines if the app is initialized
 	var isInit bool
+
+	dirCache := cache.New[[]byte](cacheStorage, serveDirchangeCacheNamespace)
 
 	// determine if the app must reset the state
 	// if the state must be reset, then we consider the chain as being not initialized
@@ -275,7 +277,7 @@ func (c *Chain) serve(ctx context.Context, forceReset bool) error {
 	if isInit {
 		configModified := false
 		if c.ConfigPath() != "" {
-			configModified, err = dirchange.HasDirChecksumChanged(c.app.Path, []string{c.ConfigPath()}, saveDir, configChecksum)
+			configModified, err = dirchange.HasDirChecksumChanged(dirCache, configChecksumKey, c.app.Path, c.ConfigPath())
 			if err != nil {
 				return err
 			}
@@ -290,7 +292,7 @@ func (c *Chain) serve(ctx context.Context, forceReset bool) error {
 
 	// check if source has been modified since last serve
 	// if the state must not be reset but the source has changed, we rebuild the chain and import the exported state
-	sourceModified, err := dirchange.HasDirChecksumChanged(c.app.Path, appBackendSourceWatchPaths, saveDir, sourceChecksum)
+	sourceModified, err := dirchange.HasDirChecksumChanged(dirCache, sourceChecksumKey, c.app.Path, appBackendSourceWatchPaths...)
 	if err != nil {
 		return err
 	}
@@ -308,7 +310,7 @@ func (c *Chain) serve(ctx context.Context, forceReset bool) error {
 		}
 		binaryModified = true
 	} else {
-		binaryModified, err = dirchange.HasDirChecksumChanged("", []string{binaryPath}, saveDir, binaryChecksum)
+		binaryModified, err = dirchange.HasDirChecksumChanged(dirCache, binaryChecksumKey, "", binaryPath)
 		if err != nil {
 			return err
 		}
@@ -331,7 +333,7 @@ func (c *Chain) serve(ctx context.Context, forceReset bool) error {
 	// build phase
 	if !isInit || appModified {
 		// build the blockchain app
-		if err := c.build(ctx, ""); err != nil {
+		if err := c.build(ctx, cacheStorage, ""); err != nil {
 			return err
 		}
 	}
@@ -362,18 +364,18 @@ func (c *Chain) serve(ctx context.Context, forceReset bool) error {
 
 	// save checksums
 	if c.ConfigPath() != "" {
-		if err := dirchange.SaveDirChecksum(c.app.Path, []string{c.ConfigPath()}, saveDir, configChecksum); err != nil {
+		if err := dirchange.SaveDirChecksum(dirCache, configChecksumKey, c.app.Path, c.ConfigPath()); err != nil {
 			return err
 		}
 	}
-	if err := dirchange.SaveDirChecksum(c.app.Path, appBackendSourceWatchPaths, saveDir, sourceChecksum); err != nil {
+	if err := dirchange.SaveDirChecksum(dirCache, sourceChecksumKey, c.app.Path, appBackendSourceWatchPaths...); err != nil {
 		return err
 	}
 	binaryPath, err = exec.LookPath(binaryName)
 	if err != nil {
 		return err
 	}
-	if err := dirchange.SaveDirChecksum("", []string{binaryPath}, saveDir, binaryChecksum); err != nil {
+	if err := dirchange.SaveDirChecksum(dirCache, binaryChecksumKey, "", binaryPath); err != nil {
 		return err
 	}