+++ ignite/services/scaffolder/query.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/gobuffalo/genny"
 
+	"github.com/ignite-hq/cli/ignite/pkg/cache"
 	"github.com/ignite-hq/cli/ignite/pkg/multiformatname"
 	"github.com/ignite-hq/cli/ignite/pkg/placeholder"
 	"github.com/ignite-hq/cli/ignite/pkg/xgenny"
@@ -16,6 +17,7 @@ import (
 // AddQuery adds a new query to scaffolded app
 func (s Scaffolder) AddQuery(
 	ctx context.Context,
+	cacheStorage cache.Storage,
 	tracer *placeholder.Tracer,
 	moduleName,
 	queryName,
@@ -68,7 +70,6 @@ func (s Scaffolder) AddQuery(
 			AppPath:     s.path,
 			ModulePath:  s.modpath.RawPath,
 			ModuleName:  moduleName,
-			OwnerName:   owner(s.modpath.RawPath),
 			QueryName:   name,
 			ReqFields:   parsedReqFields,
 			ResFields:   parsedResFields,
@@ -86,5 +87,5 @@ func (s Scaffolder) AddQuery(
 	if err != nil {
 		return sm, err
 	}
-	return sm, finish(opts.AppPath, s.modpath.RawPath)
+	return sm, finish(cacheStorage, opts.AppPath, s.modpath.RawPath)
 }