+++ integration/cosmosgen/cosmosgen_test.go
@@ -15,7 +15,7 @@ import (
 func TestCosmosGen(t *testing.T) {
 	var (
 		env          = envtest.New(t)
-		path         = env.Scaffold("blog")
+		path         = env.Scaffold("github.com/test/blog")
 		dirGenerated = filepath.Join(path, "vue/src/store/generated")
 	)
 
@@ -116,6 +116,7 @@ func TestCosmosGen(t *testing.T) {
 
 	var expectedCosmosModules = []string{
 		"cosmos.auth.v1beta1",
+		"cosmos.authz.v1beta1",
 		"cosmos.bank.v1beta1",
 		"cosmos.base.tendermint.v1beta1",
 		"cosmos.crisis.v1beta1",